CIF-1328: Moved navigation operations to actions
